(net) obsd warnings
FossilOrigin-Name: 11cc3d58296473dddd9e19e240dea051f4d49857eb52513b9df29db394e50361

+++ src/a12/net/dir_srv.c
@@ -595,7 +595,7 @@ int dirsrv_build_report(const char* appl)
 			continue;
 
 		unsigned char* b64 = a12helper_tob64(outk, 32, &(size_t){0});
-		fprintf(out, "source=%s:length=%zd\n", b64, sbuf.st_size + 1);
+		fprintf(out, "source=%s:length=%zd\n", b64, (ssize_t) sbuf.st_size + 1);
 		free(b64);
 
 		while (!feof(fin)){
